changed to only apply to dedicated assignement
other pool types shouldn't need cleaning of assignments

+++ scripts/cleanEntitlements.ps1
@@ -2,29 +2,32 @@
 
 # load necessary plugin
 . "C:\Program Files\VMware\VMware View\Server\extras\PowerShell\add-snapin.ps1"
-. "C:\Users\Public\Documents\scripts\constants.ps1"
+. "C:\path\scripts\constants.ps1"
 
 foreach ($pool in Get-Pool) {
-	# arrays
-	$remove = @()
-	$user = @()
+	# for all automated pools with dedicated assignement
+	if ($pool.poolType -eq "Persistent" -or $pool.poolType -eq "ManualVCPersistent") {
+		# arrays
+		$remove = @()
+		$user = @()
 
-	# get all users
-	foreach ($vm in Get-DesktopVM -Pool_id $pool.pool_id) {
-		if ($vm.user_displayname -ne "") {
-			$user += $vm.user_displayname
+		# get all users
+		foreach ($vm in Get-DesktopVM -Pool_id $pool.pool_id) {
+			if ($vm.user_displayname -ne "") {
+				$user += $vm.user_displayname
+			}
 		}
-	}
 
-	# check if entitlement has a user
-	foreach ($entitlement in Get-PoolEntitlement -Pool_id $pool.pool_id) {
-		if ($entitlement.displayName -notin $user) {
-			$remove += $entitlement.sid
+		# check if entitlement has a user
+		foreach ($entitlement in Get-PoolEntitlement -Pool_id $pool.pool_id) {
+			if ($entitlement.displayName -notin $user) {
+				$remove += $entitlement.sid
+			}
 		}
-	}
 
-	# remove entitlements without a user
-	foreach ($sid in $remove) {
-		Remove-PoolEntitlement -Pool_id $pool.pool_id -Sid $sid
+		# remove entitlements without a user
+		foreach ($sid in $remove) {
+			Remove-PoolEntitlement -Pool_id $pool.pool_id -Sid $sid
+		}
 	}
 }
